back/requirements.txt
aiogram>=3.13.1
python-dotenv==1.0.1
openai>=1.30.0
aiohttp>=3.9.5
pydantic>=2.10.4
uvicorn>=0.30.0
fastapi>=0.111.0

